Move models to models.py module
utils.types is meant for type aliases not models
